minor changes to particlesystem.

+++ 3DGameEngine/3DGameEngine/res/basicParticlePhy.vert
@@ -5,7 +5,7 @@ layout (location = 1) in vec3 Position;
 layout (location = 2) in vec3 Velocity;                                             
 layout (location = 3) in float Age;                                                 
                                                                                     
-out float Type0;                                                                    
+out int Type0;                                                                    
 out vec3 Position0;                                                                 
 out vec3 Velocity0;                                                                 
 out float Age0;
